feature/SCIHUB-1209 add component prop
