+++ src/lib/components/AddDocumentModal/AddDocumentModalContent.tsx
@@ -1,5 +1,4 @@
 import { AnimatePresence, motion } from 'framer-motion';
-import { useMemo } from 'react';
 import { useWatch } from 'react-hook-form';
 import styled from 'styled-components';
